bazelProcess moved to the bazel package
