+++ src/main/java/org/jetbrains/bsp/bazel/server/bazel/BazelRunnerBuilder.java
@@ -3,7 +3,6 @@ package org.jetbrains.bsp.bazel.server.bazel;
 import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
-import org.jetbrains.bsp.bazel.server.bazel.data.BazelProcess;
 import org.jetbrains.bsp.bazel.server.bazel.params.BazelQueryKindParameters;
 import org.jetbrains.bsp.bazel.server.bazel.params.BazelRunnerFlag;
 import org.jetbrains.bsp.bazel.server.bazel.utils.BazelArgumentsUtils;
